Add escaping single quotes in PostgresProcessor
TableExists, ColumnExists, ConstraintExists, SchemaExists and
IndexExists in PostgresProcessor do not accept single quotes in their
name parameters e.g. Test'Schema. This means that the MigrationRunner
integration tests do not work for Postgres. This is only an issue with
testing and is not currently something that effects the end user.

This change adds escaping for the schema name, table name, column name,
index name and constraint name parameters in the above named methods as
well as the appropriate integration tests. There are some changes to
PostgresTestTable as well to enable the passing of table names into
the constructor.